requirements.txt
# pip install -r requirements.txt

# base ----------------------------------------
Cython
matplotlib>=3.2.2
numpy>=1.18.5
opencv-python~=3.4.2.16
pillow~=7.2.0
PyYAML>=5.3
scipy>=1.4.1
tensorboard>=2.2
torch>=1.6.0
torchvision>=0.7.0
tqdm>=4.41.0

yaml~=0.2.5
labelme~=4.5.6

future==0.18.2
MarkupSafe==1.1.1
protobuf==3.11.3
pyparsing==2.4.6
python-dateutil==2.8.1
six==1.14.0
